// src/app/common/model/casualty.model.ts
export interface MonthlyCasualtyRecordModel {
    specifiedBranch: string;
    date: string;
    nameOfDeceased: string;
    address: string;
    membership: string;
    typeOfPlan: string;
    planPrice: string;
    monthlyContribution: string;
    contribution: string;
    discount: string;
    governmentAssistance: string;
    cashDeposit: string;
    paymentStatus: string;
  }
  